+++ src/modules/voxelformat/VolumeFormat.cpp
@@ -154,6 +154,7 @@ const io::FormatDescription *voxelLoad() {
 		{"Wavefront Object", {"obj"}, nullptr, VOX_FORMAT_FLAG_MESH},
 		{"GL Transmission Format", {"gltf", "glb", "vrm"}, nullptr, VOX_FORMAT_FLAG_MESH},
 		{"Standard Triangle Language", {"stl"}, nullptr, VOX_FORMAT_FLAG_MESH},
+		{"Polygon File Format", {"ply"}, nullptr, VOX_FORMAT_FLAG_MESH},
 		{"Build engine", {"kvx"}, nullptr, VOX_FORMAT_FLAG_PALETTE_EMBEDDED},
 		{"AceOfSpades",
 		 {"kv6"},